requirements.txt
Package                   Version
------------------------- -----------
altair                    6.0.0
annotated-types           0.7.0
anyio                     4.13.0
attrs                     26.1.0
blinker                   1.9.0
cachetools                7.0.5
certifi                   2026.2.25
cffi                      2.0.0
charset-normalizer        3.4.7
click                     8.1.8
colorama                  0.4.6
cryptography              46.0.7
distro                    1.9.0
gitdb                     4.0.12
GitPython                 3.1.46
google-auth               2.49.2
google-genai              1.73.0
gTTS                      2.5.4
h11                       0.16.0
httpcore                  1.0.9
httpx                     0.28.1
idna                      3.11
Jinja2                    3.1.6
jsonschema                4.26.0
jsonschema-specifications 2025.9.1
MarkupSafe                3.0.3
narwhals                  2.19.0
numpy                     2.4.4
packaging                 26.0
pandas                    3.0.2
pillow                    12.2.0
pip                       26.0.1
protobuf                  7.34.1
pyarrow                   23.0.1
pyasn1                    0.6.3
pyasn1_modules            0.4.2
pycparser                 3.0
pydantic                  2.13.0
pydantic_core             2.46.0
pydeck                    0.9.1
python-dateutil           2.9.0.post0
python-dotenv             1.2.2
referencing               0.37.0
requests                  2.33.1
rpds-py                   0.30.0
setuptools                65.5.0
six                       1.17.0
smmap                     5.0.3
sniffio                   1.3.1
streamlit                 1.56.0
tenacity                  9.1.4
toml                      0.10.2
tornado                   6.5.5
typing_extensions         4.15.0
typing-inspection         0.4.2
tzdata                    2026.1
urllib3                   2.6.3
watchdog                  6.0.0
websockets                16.0